Ported Tickets system from TC.
Closes #171

+++ src/server/game/Battlegrounds/BattleGroundMgr.h
@@ -120,14 +120,14 @@ class TC_GAME_API BattlegroundMgr
             CreatureBattleEventIndexesMap::const_iterator itr = m_CreatureBattleEventIndexMap.find(spawnID);
             if (itr != m_CreatureBattleEventIndexMap.end())
                 return itr->second[0];
-            return m_CreatureBattleEventIndexMap.find(-1)->second[0];
+            return m_CreatureBattleEventIndexMap.find(uint32(-1))->second[0];
         }
         const BattleGroundEventIdx GetGameObjectEventIndex(uint32 spawnID) const
         {
             GameObjectBattleEventIndexesMap::const_iterator itr = m_GameObjectBattleEventIndexMap.find(spawnID);
             if (itr != m_GameObjectBattleEventIndexMap.end())
                 return itr->second[0];
-            return m_GameObjectBattleEventIndexMap.find(-1)->second[0];
+            return m_GameObjectBattleEventIndexMap.find(uint32(-1))->second[0];
         }
         // Nostalrius: allow multiple events per creature ... Avoid when possible.
         std::vector<BattleGroundEventIdx> const& GetCreatureEventsVector(uint32 spawnID) const
@@ -135,14 +135,14 @@ class TC_GAME_API BattlegroundMgr
             CreatureBattleEventIndexesMap::const_iterator itr = m_CreatureBattleEventIndexMap.find(spawnID);
             if (itr != m_CreatureBattleEventIndexMap.end())
                 return itr->second;
-            return m_CreatureBattleEventIndexMap.find(-1)->second;
+            return m_CreatureBattleEventIndexMap.find(uint32(-1))->second;
         }
         std::vector<BattleGroundEventIdx> const& GetGameObjectEventsVector(uint32 spawnID) const
         {
             GameObjectBattleEventIndexesMap::const_iterator itr = m_GameObjectBattleEventIndexMap.find(spawnID);
             if (itr != m_GameObjectBattleEventIndexMap.end())
                 return itr->second;
-            return m_GameObjectBattleEventIndexMap.find(-1)->second;
+            return m_GameObjectBattleEventIndexMap.find(uint32(-1))->second;
         }
 
         void InitAutomaticArenaPointDistribution();
